pyproject.toml
[project]
name = "jax_rl"
version = "0.1.0"
description = "Pure JAX Proximal Policy Optimization baseline"
requires-python = ">=3.11"
dependencies = [
    "jax[cuda12]>=0.5.0",
    "jaxlib>=0.5.0",
    "flax>=0.11.0",
    "orbax-checkpoint>=0.11.0",
    "optax>=0.2.4",
    "distrax>=0.1.5",
    "pyyaml>=6.0.0",
    "gymnax>=0.0.8",
    "stoa-env",
    "rustpool",
    "rlpallet",
    "jaxpallet",
    "tensorboard>=2.6",
    "colorama>=0.4.6",
    "setuptools==80.9.0",
    "hydra-core>=1.3.2",
    "pytest>=8.0.0",
    "jax2onnx",
    "onnx",
    "onnxruntime",
    "nvitop",
    "mctx",
    "flashbax",
]

[dependency-groups]
dev = [
    "pyright>=1.1.390",
]

[project.scripts]
jax-rl-train = "jax_rl.cli:main"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = ["-ra", "--strict-config", "--strict-markers"]
markers = [
    "integration: slower end-to-end tests that exercise train/evaluator loops",
]
filterwarnings = [
    'ignore:Setting `jax_pmap_shmap_merge` is deprecated in JAX v0\.9\.0 and will be removed in JAX v0\.10\.0\.:DeprecationWarning:jax\._src\.config',
    'ignore:Sharding info not provided when restoring\.:UserWarning:orbax\.checkpoint\._src\.serialization\.jax_array_handlers',
    'ignore:jax\.core\.get_aval is deprecated; use jax\.typeof instead\.:DeprecationWarning:distrax\._src\.utils\.math',
    "ignore:'\\.value' access is now deprecated\\.:DeprecationWarning",
]

[tool.uv.sources]
stoa-env = { git = "https://github.com/EdanToledo/Stoa.git", rev = "main" }
rustpool = { path = "../rustpool" }
rlpallet = { path = "../rlpallet", editable = true }
jaxpallet = { path = "../jaxpallet", editable = true }

[tool.pyright]
include = ["src/jax_rl"]
exclude = ["tests"]
typeCheckingMode = "standard"
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false
